🚤🐾 Updated with Glitch

+++ komutlar/oylama.js
@@ -1,48 +1,25 @@
 const Discord = require('discord.js');
 
- exports.run = (client, message, args) => {
-   message.delete();
+exports.run = (client, message, args) => {
+	let mesaj = args.slice(0).join(' ');
+	if (mesaj.length < 1) return message.reply('Yazmam İçin Birşey Yazmalısın!');
+    message.delete();
+    const embed = new Discord.RichEmbed()
+    .setAuthor('OYLAMA')
+    .setColor(3447003)
+    .setDescription(`${mesaj} \n\n\ Evet İçin: :thumbsup: Hayır İçin: :thumbsdown: `)
+    return message.channel.sendEmbed(embed);
+};
 
-   let question = args.join(' ');
-
-   let user = message.author.username
-
-   if (!question) return message.channel.sendEmbed(
-
-     new Discord.RichEmbed()
-
-     .addField(`:x:yazı yazman gerek :x:`)).then(m => m.delete(5000));
-
-     console.log("/oylama komutu " + message.author.username + '#' + message.author.discriminator + " tarafından kullanıldı.")
-     message.channel.sendEmbed(
-
-       new Discord.RichEmbed()
-
-       .setColor("RANDOM")
-       .setThumbnail(client.user.avatarURL)
-       .setTimestamp()
-       .setFooter('MeeTR', client.user.avatarURL)
-
-       .addField(`**Oylama**`, `**${question}**`)).then(function(message) {
-
-         message.react('✅');
-
-         message.react('❌');
-
-       });
-
-     };
-
-     exports.conf = {
-       enabled: true,
-       guildOnly: false,
-       aliases: ['oylama'],
-
+exports.conf = {
+  enabled: true,
+  guildOnly: false,
+  aliases: [],
   permLevel: 2
 };
 
 exports.help = {
   name: 'oylama',
-  description: 'Oylama yapmanızı sağlar.',
-  usage: '/oylama <oylamaismi>'
+  description: 'Oylama Yapar.',
+  usage: 'oylama'
 };
